requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
Faker>=20.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1

